Player Movement and Interaction Screen
Made a movable player, He can jump, air dash sprint and wall jump. Also made a dynamic camera that moves closer to the player if it collides with an object.

Also made a script that allows the player to "interact" with an object. This stops play movement and moves the camera to view the object. We can use this to lock in on the pinball machine when playing.
Controls:
wasd movement
shift (on ground) - sprint
shift (in air) - dash
space - jump (or air jump)
mouse controls camera, press esc to free mouse from locked

E to interact and back out of interaction